optimize expire snapshots fix: logical error remove some comments

+++ amoro-format-iceberg/src/main/java/org/apache/amoro/utils/TableFileUtil.java
@@ -28,6 +28,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.net.URI;
+import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -75,15 +76,24 @@ public class TableFileUtil {
     return filePath.substring(0, lastSlash);
   }
 
+  public static void deleteEmptyDirectory(
+      AuthenticatedFileIO io, String directoryPath, Set<String> exclude) {
+    deleteEmptyDirectory(io, directoryPath, exclude, new HashSet<>());
+  }
+
   /**
    * Try to recursiveDelete the empty directory
    *
    * @param io mixed-format file io
    * @param directoryPath directory location
-   * @param exclude the directory will not be deleted
+   * @param exclude the directory will not be deleted @Param directoriesToBeDeleted: all the
+   *     directories that need to be deleted
    */
   public static void deleteEmptyDirectory(
-      AuthenticatedFileIO io, String directoryPath, Set<String> exclude) {
+      AuthenticatedFileIO io,
+      String directoryPath,
+      Set<String> exclude,
+      Set<String> directoriesToBeDeleted) {
     if (directoryPath == null || directoryPath.isEmpty()) {
       return;
     }
@@ -108,7 +118,10 @@ public class TableFileUtil {
     if (io.asFileSystemIO().isEmptyDirectory(directoryPath)) {
       io.asFileSystemIO().deletePrefix(directoryPath);
       LOG.debug("success delete empty directory {}", directoryPath);
-      deleteEmptyDirectory(io, parent, exclude);
+      // for parent must be deleted after the sub-directory
+      if (directoriesToBeDeleted == null || !directoriesToBeDeleted.contains(parent)) {
+        deleteEmptyDirectory(io, parent, exclude);
+      }
     }
   }
